# mock-server/openapi.yaml
openapi: 3.0.0
info:
  title: Smart Clinic OS API Specification (SCS)
  description: |
    المستند البرمجي لواجهات برمجة التطبيقات (API Specification) لمنصة عيادتي الذكية.
    يحتوي الملف على مسارين أساسيين:
    1. **واجهات الأوبريشن والمشغل (Platform Operations):** تبدأ بـ `/admin/v1` ومحمية بـ VPN.
    2. **واجهات العيادات والعملاء (Clinic Dashboard):** تبدأ بـ `/v1` وتعتمد على ترويسة العيادة.
  version: 1.0.0
servers:
  - url: https://api.SCS-admin.com
    description: Public Clinic Dashboard Gateway
  - url: https://api.SCS-ops.com
    description: Private Operations Gateway (Accessible via VPN only)
  - url: http://localhost:3000
    description: Local Mock Development Server
paths:
...
  /v1/chats/{conversation_id}/toggle-bot:
    post:
      summary: تفعيل/تعطيل البوت يدوياً (Toggle Chat Bot Manual Takeover)
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - manual_mode
              properties:
                manual_mode:
                  type: boolean
                  example: true
      responses:
        '200':
          description: تم تبديل التحكم بنجاح
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      conversation_id:
                        type: string
                      bot_active:
                        type: boolean
                        example: false
                      manual_mode_expires_at:
                        type: string
                        example: "2026-07-01T23:00:00Z"
  /v1/appointments/{appointment_id}/consultation:
    post:
      summary: كتابة التشخيص واعتماد الروشتة (E-Prescription & SOAP Notes)
      parameters:
        - name: appointment_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - subjective
                - objective
                - diagnosis_icd11
                - prescription_items
              properties:
                subjective:
                  type: string
                  example: يعاني المريض من ألم شديد في الضرس الخلفي السفلي
                objective:
                  type: object
                  properties:
                    blood_pressure:
                      type: string
                      example: 120/80
                    pulse:
                      type: integer
                      example: 72
                diagnosis_icd11:
                  type: string
                  example: DA01.1 - Dental caries extending into dentine
                prescription_items:
                  type: array
                  items:
                    type: object
                    properties:
                      medication_name:
                        type: string
                        example: Amoxicillin 500mg
                      dosage:
                        type: string
                        example: كبسولة واحدة كل 8 ساعات
                      duration:
                        type: string
                        example: 5 أيام
      responses:
        '200':
          description: تم اعتماد الزيارة وتوليد الروشتة وإرسالها
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      medical_record_id:
                        type: string
                        example: record-uuid-11111
                      prescription_id:
                        type: string
                        example: prescription-uuid-22222
                      pdf_url:
                        type: string
                        example: https://storage.SCS-admin.com/tenant-id/prescriptions/rx-2026-0001.pdf
                      whatsapp_status:
                        type: string
                        example: enqueued
...